Ide coccoa proj fixes.

+++ Ide/Code/FindSymbol.cpp
@@ -402,7 +402,7 @@ struct FindSymbolSystemPriv : public GEventTargetThread
 				{
 					MsgTs = Now;
 					DoingProgress = true;
-					uint32 Remaining = Tasks - GetQueueSize();
+					uint32 Remaining = (uint32) (Tasks - GetQueueSize());
 					if (Remaining > 0)
 						Log("FindSym: %i of %i (%.1f%%)\n", Remaining, Tasks, (double)Remaining * 100.0 / MAX(Tasks, 1));
 				}
